Force `rejected_events_metadata` background update to run in foreground
