+++ synapse/storage/databases/main/events_bg_updates.py
@@ -37,7 +37,7 @@ from synapse.crypto.event_signing import (
     event_needs_resigning,
     resign_event,
 )
-from synapse.events import EventBase, make_event_from_dict
+from synapse.events import EventBase
 from synapse.storage._base import SQLBaseStore, db_to_json, make_in_list_sql_clause
 from synapse.storage.database import (
     DatabasePool,
@@ -205,11 +205,6 @@ class EventsBackgroundUpdatesStore(
             index_name="users_have_local_media",
             table="local_media_repository",
             columns=["user_id", "created_ts"],
-        )
-
-        self.db_pool.updates.register_background_update_handler(
-            "rejected_events_metadata",
-            self._rejected_events_metadata,
         )
 
         self.db_pool.updates.register_background_update_handler(
@@ -922,123 +917,6 @@ class EventsBackgroundUpdatesStore(
             await self.db_pool.updates._end_background_update("event_store_labels")
 
         return num_rows
-
-    async def _rejected_events_metadata(self, progress: dict, batch_size: int) -> int:
-        """Adds rejected events to the `state_events` and `event_auth` metadata
-        tables.
-        """
-
-        last_event_id = progress.get("last_event_id", "")
-
-        def get_rejected_events(
-            txn: Cursor,
-        ) -> list[tuple[str, str, JsonDict, bool, bool]]:
-            # Fetch rejected event json, their room version and whether we have
-            # inserted them into the state_events or auth_events tables.
-            #
-            # Note we can assume that events that don't have a corresponding
-            # room version are V1 rooms.
-            sql = """
-                SELECT DISTINCT
-                    event_id,
-                    COALESCE(room_version, '1'),
-                    json,
-                    state_events.event_id IS NOT NULL,
-                    event_auth.event_id IS NOT NULL
-                FROM rejections
-                INNER JOIN event_json USING (event_id)
-                LEFT JOIN rooms USING (room_id)
-                LEFT JOIN state_events USING (event_id)
-                LEFT JOIN event_auth USING (event_id)
-                WHERE event_id > ?
-                ORDER BY event_id
-                LIMIT ?
-            """
-
-            txn.execute(
-                sql,
-                (
-                    last_event_id,
-                    batch_size,
-                ),
-            )
-
-            return cast(
-                list[tuple[str, str, JsonDict, bool, bool]],
-                [(row[0], row[1], db_to_json(row[2]), row[3], row[4]) for row in txn],
-            )
-
-        results = await self.db_pool.runInteraction(
-            desc="_rejected_events_metadata_get", func=get_rejected_events
-        )
-
-        if not results:
-            await self.db_pool.updates._end_background_update(
-                "rejected_events_metadata"
-            )
-            return 0
-
-        state_events = []
-        auth_events = []
-        for event_id, room_version, event_json, has_state, has_event_auth in results:
-            last_event_id = event_id
-
-            if has_state and has_event_auth:
-                continue
-
-            room_version_obj = KNOWN_ROOM_VERSIONS.get(room_version)
-            if not room_version_obj:
-                # We no longer support this room version, so we just ignore the
-                # events entirely.
-                logger.info(
-                    "Ignoring event with unknown room version %r: %r",
-                    room_version,
-                    event_id,
-                )
-                continue
-
-            event = make_event_from_dict(event_json, room_version_obj)
-
-            if not event.is_state():
-                continue
-
-            if not has_state:
-                state_events.append(
-                    (event.event_id, event.room_id, event.type, event.state_key)
-                )
-
-            if not has_event_auth:
-                # Old, dodgy, events may have duplicate auth events, which we
-                # need to deduplicate as we have a unique constraint.
-                for auth_id in set(event.auth_event_ids()):
-                    auth_events.append((event.event_id, event.room_id, auth_id))
-
-        if state_events:
-            await self.db_pool.simple_insert_many(
-                table="state_events",
-                keys=("event_id", "room_id", "type", "state_key"),
-                values=state_events,
-                desc="_rejected_events_metadata_state_events",
-            )
-
-        if auth_events:
-            await self.db_pool.simple_insert_many(
-                table="event_auth",
-                keys=("event_id", "room_id", "auth_id"),
-                values=auth_events,
-                desc="_rejected_events_metadata_event_auth",
-            )
-
-        await self.db_pool.updates._background_update_progress(
-            "rejected_events_metadata", {"last_event_id": last_event_id}
-        )
-
-        if len(results) < batch_size:
-            await self.db_pool.updates._end_background_update(
-                "rejected_events_metadata"
-            )
-
-        return len(results)
 
     async def _chain_cover_index(self, progress: dict, batch_size: int) -> int:
         """A background updates that iterates over all rooms and generates the
